+++ ElevatorSystem.Admin/ElevatorSystem.Admin/Controllers/AdminController/ElevatorsController.cs
@@ -19,6 +19,11 @@ namespace ElevatorSystem.Admin.Controllers.AdminController
         public ActionResult Index()
         {
             var elevators = db.Elevators.Include(e => e.Category);
+            TempData["0"] = "Status : 0 - Pending";
+            TempData["1"] = "Status : 1 - Available";
+            TempData["2"] = "Status : 2 - Upgrading";
+            TempData["3"] = "Status : 3 - Out of date";
+            TempData["n"] = "Status : n - Status is undefined";
             return View(elevators.ToList());
         }
 
@@ -30,6 +35,26 @@ namespace ElevatorSystem.Admin.Controllers.AdminController
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             }
             Elevator elevator = db.Elevators.Find(id);
+            if (elevator.Status == 0)
+            {
+                ViewData["elevatorStatus"] = "Pending";
+            }
+            else if (elevator.Status == 1)
+            {
+                ViewData["elevatorStatus"] = "Available";
+            }
+            else if (elevator.Status == 2)
+            {
+                ViewData["elevatorStatus"] = "Upgrading";
+            }
+            else if (elevator.Status == 3)
+            {
+                ViewData["elevatorStatus"] = "Out of date";
+            }
+            else
+            {
+                ViewData["elevatorStatus"] = "Status is undefined";
+            }
             if (elevator == null)
             {
                 return HttpNotFound();
@@ -40,33 +65,48 @@ namespace ElevatorSystem.Admin.Controllers.AdminController
         // GET: Elevators/Create
         public ActionResult Create()
         {
-           // ViewBag.CategoryID = new SelectList(db.Categories, "ID", "Name");
-            return View("Create");
-        }
-
-        [HttpPost]
-        public string UploadImages(HttpPostedFileBase file)
-        {
-            Random r = new Random();
-            int num = r.Next();
-
-            file.SaveAs(Server.MapPath("~/Content/Elevator/") + num + "_" + file.FileName);
-            return "/Content/Elevator/" + num + "_" + file.FileName;
+            
+            ViewBag.CategoryID = new SelectList(db.Categories, "ID", "Name");
+            return View();
         }
 
         // POST: Elevators/Create
         // To protect from overposting attacks, enable the specific properties you want to bind to, for 
         // more details see https://go.microsoft.com/fwlink/?LinkId=317598.
+        /* [HttpPost]
+         [ValidateAntiForgeryToken]
+         public ActionResult Create([Bind(Include = "ID,Name,SKU,Status,Description,Thumbnails,Capacity,Speed,Price,MaxPerson,Location,Slug,Tag,CreatedAt,UpdatedAt,DeletedAt,CategoryID")] Elevator elevator)
+         {
+
+             if (ModelState.IsValid)
+             {
+                 Random rd = new Random();
+                 elevator.SKU = "ELEVATOR00" + rd.Next(1,1000).ToString();
+                 elevator.CreatedAt = DateTime.Today;
+                 db.Elevators.Add(elevator);
+                 db.SaveChanges();
+                 TempData["CreateMessage"] = "Elevator { #" + elevator.ID + "." + elevator.Name + " } has been added to the list !";
+                 return RedirectToAction("Index");
+             }
+
+             ViewBag.CategoryID = new SelectList(db.Categories, "ID", "Name", elevator.CategoryID);
+             return View(elevator);
+         }*/
+        
         [HttpPost]
-        public JsonResult Create( Elevator elevator)
+        public JsonResult Create(Elevator elevator)
         {
+
             if (ModelState.IsValid)
             {
+                Random rd = new Random();
+                elevator.SKU = "ELEVATOR00" + rd.Next(1, 1000).ToString();
+                elevator.CreatedAt = DateTime.Now;
                 db.Elevators.Add(elevator);
                 db.SaveChanges();
+                return Json(elevator, JsonRequestBehavior.AllowGet);
             }
-
-           // ViewBag.CategoryID = new SelectList(db.Categories, "ID", "Name", elevator.CategoryID);
+           
             return Json(elevator, JsonRequestBehavior.AllowGet);
         }
 
@@ -91,15 +131,18 @@ namespace ElevatorSystem.Admin.Controllers.AdminController
         // more details see https://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult Edit(Elevator elevator)
+        public ActionResult Edit([Bind(Include = "ID,Name,SKU,Status,Description,Thumbnails,Capacity,Speed,Price,MaxPerson,Location,Slug,Tag,CreatedAt,UpdatedAt,DeletedAt,CategoryID")] Elevator elevator)
         {
             if (ModelState.IsValid)
             {
+                elevator.UpdatedAt = DateTime.Today;
+                
                 db.Entry(elevator).State = EntityState.Modified;
                 db.SaveChanges();
+                TempData["UpdateMessage"] = "Elevator { #" + elevator.ID + "." + elevator.Name + " } has been updated !";
                 return RedirectToAction("Index");
             }
-           // ViewBag.CategoryID = new SelectList(db.Categories, "ID", "Name", elevator.CategoryID);
+            ViewBag.CategoryID = new SelectList(db.Categories, "ID", "Name", elevator.CategoryID);
             return View(elevator);
         }
 
@@ -111,6 +154,28 @@ namespace ElevatorSystem.Admin.Controllers.AdminController
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             }
             Elevator elevator = db.Elevators.Find(id);
+            if (elevator.Status == 0)
+            {
+                ViewData["elevatorStatus"] = "Pending";
+            }
+            else if (elevator.Status == 1)
+            {
+                ViewData["elevatorStatus"] = "Available";
+            }
+            else if (elevator.Status == 2)
+            {
+                ViewData["elevatorStatus"] = "Upgrading";
+            }
+            else if (elevator.Status == 3)
+            {
+                ViewData["elevatorStatus"] = "Out of date";
+            }
+            else
+            {
+                ViewData["elevatorStatus"] = "Status is undefined";
+            }
+
+
             if (elevator == null)
             {
                 return HttpNotFound();
@@ -124,9 +189,51 @@ namespace ElevatorSystem.Admin.Controllers.AdminController
         public ActionResult DeleteConfirmed(int id)
         {
             Elevator elevator = db.Elevators.Find(id);
-            db.Elevators.Remove(elevator);
-            db.SaveChanges();
-            return RedirectToAction("Index");
+            //Check Status
+            if (elevator.Status == 0)
+            {
+                ViewData["elevatorStatus"] = "Pending";
+            }
+            else if (elevator.Status == 1)
+            {
+                ViewData["elevatorStatus"] = "Available";
+            }
+            else if (elevator.Status == 2)
+            {
+                ViewData["elevatorStatus"] = "Upgrading";
+            }
+            else if (elevator.Status == 3)
+            {
+                ViewData["elevatorStatus"] = "Out of date";
+            }
+            else
+            {
+                ViewData["elevatorStatus"] = "Status is undefined";
+            }
+
+            //Check Valid Delete
+            if (elevator.Status >= 3)
+            {
+                db.Elevators.Remove(elevator);
+                db.SaveChanges();
+                TempData["DeleteMessage"] = "Elevator { #" + elevator.ID + "." + elevator.Name + " } has been removed from the list !";
+                return RedirectToAction("Index");
+            }
+            else
+            {
+                ViewData["InvalidStatus"] = "You are not allowed to remove Elevator { #" + elevator.ID + "." + elevator.Name + " } !";
+                return View(elevator);
+            }
+          
+        }
+        [HttpPost]
+        public string UploadImages(HttpPostedFileBase file)
+        {
+            Random r = new Random();
+            int num = r.Next();
+
+            file.SaveAs(Server.MapPath("~/Content/Elevator/") + num + "_" + file.FileName);
+            return "/Content/Elevator/" + num + "_" + file.FileName;
         }
 
         protected override void Dispose(bool disposing)
